1626. Best Team With No Conflicts
